moved tests to new module
